Move callback files to callback directory
