+++ smac/callback/__init__.py
@@ -1,0 +1,7 @@
+from smac.callback.callback import Callback
+from smac.callback.metadata_callback import MetadataCallback
+
+__all__ = [
+    "Callback",
+    "MetadataCallback",
+]
